requirements.txt
# Essential Core
langgraph
langchain
langchain-core
langchain-openai
langsmith
openai

# Web Framework
streamlit

# APIs and Web
requests
python-dotenv

# Audio
gTTS

# Data Processing
lxml
typing-extensions

# Search APIs 
tavily-python
google-api-python-client

# Visualization
ipython
graphviz

# Standard library alternatives (usually built-in)
 datetime, json, os, sys, re, urllib, xml, io, time, zipfile, pathlib, PyMuPDF

